remove dxtbx array_index dependency
Weirdly, dxtbx always assumes that the first image number is 1. DIALS handles cases where it starts with zero by setting a "bacth_offset" to -1. Rather than account for this, I'll just assume that the first image corresponds to the starting angle (should always be true?).

See the following discussion for details:

https://github.com/cctbx/dxtbx/issues/186