masking bug fix
The intent of the _masked  function is to apply the same mask to many different arrays, so that when compressed they have the same length. In certain circumstances, function is called with a masked array as an argument, causing it to combine the two masks. This was solved using the np.ma.getdata() function, which returns an unmasked array from a regular or masked array argument.